requirements.txt
cryptography==44.0.0
Flask==3.1.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
requests==2.32.3